[Form] Move types to namespaces
