--- packages/mui-base/src/Field/Description/FieldDescription.types.tsx
@@ -1,7 +1,0 @@
-import type { BaseUIComponentProps } from '../../utils/types';
-import type { FieldRootOwnerState } from '../Root/FieldRoot.types';
-
-export type FieldDescriptionOwnerState = FieldRootOwnerState;
-
-export interface FieldDescriptionProps
-  extends BaseUIComponentProps<'p', FieldDescriptionOwnerState> {}
