+++ packages/mui-base/src/Field/Error/FieldError.tsx
@@ -2,11 +2,12 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import { useComponentRenderer } from '../../utils/useComponentRenderer';
-import type { FieldErrorProps } from './FieldError.types';
+import { FieldRoot } from '../Root/FieldRoot';
 import { useFieldRootContext } from '../Root/FieldRootContext';
 import { useFieldError } from './useFieldError';
 import { STYLE_HOOK_MAPPING } from '../utils/constants';
 import { useFormRootContext } from '../../Form/Root/FormRootContext';
+import type { BaseUIComponentProps } from '../../utils/types';
 
 /**
  * Displays error messages for the field's control.
@@ -20,7 +21,7 @@ import { useFormRootContext } from '../../Form/Root/FormRootContext';
  * - [FieldError API](https://base-ui.netlify.app/components/react-field/#api-reference-FieldError)
  */
 const FieldError = React.forwardRef(function FieldError(
-  props: FieldErrorProps,
+  props: FieldError.Props,
   forwardedRef: React.ForwardedRef<HTMLDivElement>,
 ) {
   const { render, id, className, show, forceShow, ...otherProps } = props;
@@ -58,6 +59,22 @@ const FieldError = React.forwardRef(function FieldError(
 
   return renderElement();
 });
+
+namespace FieldError {
+  export type OwnerState = FieldRoot.OwnerState;
+
+  export interface Props extends BaseUIComponentProps<'div', OwnerState> {
+    /**
+     * Determines whether the error message should be shown when it matches a given property of the
+     * field's `ValidityState`.
+     */
+    show?: keyof ValidityState;
+    /**
+     * Determines whether the error message should be shown regardless of the field's client validity.
+     */
+    forceShow?: boolean;
+  }
+}
 
 FieldError.propTypes /* remove-proptypes */ = {
   // ┌────────────────────────────── Warning ──────────────────────────────┐
