--- packages/mui-base/src/Field/Label/FieldLabel.types.ts
@@ -1,6 +1,0 @@
-import type { BaseUIComponentProps } from '../../utils/types';
-import type { FieldRootOwnerState } from '../Root/FieldRoot.types';
-
-export type FieldLabelOwnerState = FieldRootOwnerState;
-
-export interface FieldLabelProps extends BaseUIComponentProps<'div', FieldLabelOwnerState> {}
