--- packages/mui-base/src/Field/Validity/FieldValidity.types.ts
@@ -1,12 +1,0 @@
-import type * as React from 'react';
-import type { FieldValidityData } from '../Root/FieldRoot.types';
-
-export interface FieldValidityState extends Omit<FieldValidityData, 'state'> {
-  validity: FieldValidityData['state'];
-}
-
-export interface FieldValidityOwnerState {}
-
-export interface FieldValidityProps {
-  children: (state: FieldValidityState) => React.ReactNode;
-}
