+++ packages/mui-base/src/Field/index.barrel.ts
@@ -5,9 +5,4 @@ export * from './Description/FieldDescription';
 export * from './Control/FieldControl';
 export * from './Validity/FieldValidity';
 
-export type * from './Root/FieldRoot.types';
-export type * from './Label/FieldLabel.types';
-export type * from './Error/FieldError.types';
-export type * from './Description/FieldDescription.types';
-export type * from './Control/FieldControl.types';
-export type * from './Validity/FieldValidity.types';
+export type { FieldValidityData } from './Root/FieldRoot';
