+++ packages/mui-base/src/Field/index.ts
@@ -5,25 +5,4 @@ export { FieldDescription as Description } from './Description/FieldDescription'
 export { FieldControl as Control } from './Control/FieldControl';
 export { FieldValidity as Validity } from './Validity/FieldValidity';
 
-export type {
-  FieldRootProps as RootProps,
-  FieldRootOwnerState as RootOwnerState,
-  FieldValidityData as ValidityData,
-} from './Root/FieldRoot.types';
-export type {
-  FieldLabelProps as LabelProps,
-  FieldLabelOwnerState as LabelOwnerState,
-} from './Label/FieldLabel.types';
-export type {
-  FieldErrorProps as ErrorProps,
-  FieldErrorOwnerState as ErrorOwnerState,
-} from './Error/FieldError.types';
-export type {
-  FieldDescriptionProps as DescriptionProps,
-  FieldDescriptionOwnerState as DescriptionOwnerState,
-} from './Description/FieldDescription.types';
-export type {
-  FieldControlProps as ControlProps,
-  FieldControlOwnerState as ControlOwnerState,
-} from './Control/FieldControl.types';
-export type { FieldValidityProps as ValidityProps } from './Validity/FieldValidity.types';
+export type { FieldValidityData as ValidityData } from './Root/FieldRoot';
